+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/DebuggingEnsoTest.java
@@ -92,6 +92,7 @@ public class DebuggingEnsoTest {
     engine.close();
     engine = null;
     debugger = null;
+    out.close();
   }
 
   @Before